// FrontEnd/app/types/interview.ts
export interface Interview {
  InterviewID: number
  InterviewEntryDate: string | null
  RecruiterID: number | null
  Technology: string | null
  InterviewDate: string | null
  InterviewStartTime: string | null
  InterviewMethod: string | null
  ClientID: number | null
  InterviewType: string | null
  InterviewStatus: string | null
  InterviewSupport: string | null
  InterviewFeedback: string | null
  Comments: string | null
  CreatedDate: string | null
  ModifiedDate: string | null
  EndClientName: string | null
  InterviewEndTime: string | null
}

export interface InterviewFormData {
  InterviewEntryDate: string | null
  RecruiterID: number | null
  Technology: string | null
  InterviewDate: string | null
  InterviewStartTime: string | null
  InterviewMethod: string | null
  ClientID: number | null
  InterviewType: string | null
  InterviewStatus: string | null
  InterviewSupport: string | null
  InterviewFeedback: string | null
  Comments: string | null
  EndClientName: string | null
  InterviewEndTime: string | null
}